Update Storage interface and affected files
Modify the Storage interface in srv/service.go (which is embedded in the Service interface) to add workspaceId parameter to MGet and MSet methods. Update the method signatures to include workspaceId. Update all existing calls to MGet and MSet throughout the codebase to include the workspaceId parameter and adjust the keys to ensure workspaceId is not included in the key prefix. Search for 'workspaceId', 'WorkspaceId', 'MGet', or 'MSet' in the following files to find the lines that need to be edited:
- persisted_ai/openai_activities.go
- persisted_ai/rag_activities.go
- persisted_ai/vector_activities.go
- srv/delegator.go
- coding/tree_sitter/tree_sitter_activities.go
- persisted_ai/openai_activities_test.go

Also, update the redis implementation in srv/redis/storage.go to match the new MGet and MSet method signatures, incorporating workspaceId in the key within the method implementation.