handle forced tool call in content
o3-mini has a tendency to use the content instead of tool call when
forced to use a single specific tool. this workaround handles that
situation by treating the content as if it was a tool call, in the case
that it parses as json and we forced a single tool to be used.