fix combinatorial explosion in matches
in my example class, just adding a few more functions to a class took
matches from 8 to 36 to 225, etc, with captures scaling fast as well,
and times reaching over a few minutes instead of the usual <<1s.

this fixes it by not using the alternates syntax with many alternates,
instead restructuring to use separate queries for functions within
classes.