Add a failing test for create-then-update scenario
In `dev/apply_edit_blocks_test.go`, add a new test case named "Scenario C - New File without Visible Ranges" to the `TestApplyEditBlocks_SequentialEditsSameFile` function. This test will be similar to "Scenario A", but the second `EditBlock` (the update) should be modified to have an empty `VisibleFileRanges` field. This test is expected to fail with a "no good match found" error.